insurance_calculator_app/src/main/java/lv/javaguru/travel/insurance/core/underwriting/TravelPremiumUnderwritingImpl.java
package lv.javaguru.travel.insurance.core.underwriting;

import lv.javaguru.travel.insurance.core.api.dto.AgreementDTO;
import lv.javaguru.travel.insurance.core.api.dto.PersonDTO;
import lv.javaguru.travel.insurance.core.api.dto.RiskDTO;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Component;

import java.math.BigDecimal;
import java.math.RoundingMode;
import java.util.List;

@Component
class TravelPremiumUnderwritingImpl implements TravelPremiumUnderwriting {
    @Autowired
    private SelectedRisksPremiumCalculator selectedRisksPremiumCalculator;
    @Override
    public TravelPremiumCalculationResult calculatePremium(AgreementDTO agreement, PersonDTO person){
        List<RiskDTO> travelRisks= selectedRisksPremiumCalculator.calculateSelectedRisksPremium(agreement, person);
        BigDecimal totalPremium = travelRisks.stream()
                .map(RiskDTO::getPremium)
                .reduce(BigDecimal.ZERO, BigDecimal::add)
                .setScale(2, RoundingMode.HALF_UP)
                .stripTrailingZeros();
        return new TravelPremiumCalculationResult(totalPremium,travelRisks);
    }

}
